DB is deployed, data is seeded, queries tested
