+++ client/config.js
@@ -1,4 +1,5 @@
 module.exports = {
-  aws: "ec2-18-223-158-147.us-east-2.compute.amazonaws.com"
-  // aws: "localhost:3010"
-}
+	// aws: "ec2-18-223-158-147.us-east-2.compute.amazonaws.com"
+	aws: 'ec2-184-72-197-189.compute-1.amazonaws.com'
+	// aws: "localhost:3010"
+};
