+++ proxy/public/index.html
@@ -15,6 +15,8 @@
       crossorigin="anonymous"
     />
     <link rel="stylesheet" href="http://localhost:3010/style.css">
+    <!-- <link rel="stylesheet" href="http://localhost:3010/style.css"> grahn -->
+    <!-- <link rel="stylesheet" href="http://localhost:3010/style.css"> jun -->
     <title>FEC Demo</title>
   </head>
   <body>
