+++ server/index.js
@@ -1,10 +1,11 @@
-const sqlite3 = require('sqlite3');
+// const sqlite3 = require('sqlite3');
+// const axios = require('axios');
 const express = require('express');
-const axios = require('axios');
 const parser = require('body-parser');
 const cors = require('cors');
 const path = require('path');
 const app = express();
+const pg = require('../database/pg_index.js');
 
 app.use(cors());
 app.use(express.static(__dirname + '/../client/dist'));
@@ -20,59 +21,85 @@ app.get('/:id', (req, res) => {
 	res.sendFile(path.join(__dirname + '/../client/dist/index.html'));
 });
 
-// db connect
-const db = new sqlite3.Database('./database/productHeader.db', err => {
-	if (err) console.log(err);
-	else console.log('db connect success');
-});
-
-// query DB
 app.get('/api/items/:id', (req, res) => {
-	// sends client product table
-	db.get('SELECT * FROM product WHERE id=(?)', [req.params.id], (err, response) => {
-		if (err) console.log('db get request failed:', err);
-		else {
-			res.send(response);
+	pg.connect((err, client, done) => {
+		if (err) {
+			console.log('err on pool connection in DB', err);
+		} else {
+			client.query(`SELECT * FROM product WHERE id = $1`, [req.params.id], (err, response) => {
+				if (err) {
+					console.log('error on query ', err);
+				} else {
+					res.send(response.rows);
+				}
+			});
 		}
 	});
 });
 
 app.get('/api/images/:id', (req, res) => {
-	db.all('SELECT * FROM img WHERE id_product=(?)', [req.params.id], (err, response) => {
-		if (err) console.log('db get img request failed:', err);
-		else {
-			res.send(response);
+	pg.connect((err, client, done) => {
+		if (err) {
+			console.log('err on pool connection in DB', err);
+		} else {
+			client.query(`SELECT * FROM image WHERE id_product = $1`, [req.params.id], (err, response) => {
+				if (err) {
+					console.log('error on query ', err);
+				} else {
+					res.send(response.rows);
+				}
+			});
 		}
 	});
 });
 
 app.get('/api/description/:id', (req, res) => {
-	db.all('SELECT * FROM description WHERE id_product=(?)', [req.params.id], (err, response) => {
-		if (err) console.log('db get description request failed:', err);
-		else {
-			res.send(response);
+	pg.connect((err, client, done) => {
+		if (err) {
+			console.log('err on pool connection in DB', err);
+		} else {
+			client.query(`SELECT * FROM description WHERE id_product = $1`, [req.params.id], (err, response) => {
+				if (err) {
+					console.log('error on query ', err);
+				} else {
+					res.send(response.rows);
+				}
+			});
 		}
 	});
 });
 
 app.get('/api/category/:id', (req, res) => {
-	db.all('SELECT * FROM category WHERE id_product=(?)', [req.params.id], (err, response) => {
-		if (err) console.log('db get category request failed:', err);
-		else {
-			res.send(response);
+	pg.connect((err, client, done) => {
+		if (err) {
+			console.log('err on pool connection in DB', err);
+		} else {
+			client.query(`SELECT * FROM category WHERE id_product = $1`, [req.params.id], (err, response) => {
+				if (err) {
+					console.log('error on query ', err);
+				} else {
+					res.send(response.rows);
+				}
+			});
 		}
 	});
 });
 
 app.get('/api/option_categories/:id', (req, res) => {
-	db.all('SELECT * FROM optionCategories WHERE id_category=(?)', [req.params.id], (err, response) => {
-		if (err) console.log('db get optionCategories request failed:', err);
-		else {
-			res.send(response);
+	pg.connect((err, client, done) => {
+		if (err) {
+			console.log('err on pool connection in DB', err);
+		} else {
+			client.query(`SELECT * FROM category WHERE id_category = $1`, [req.params.id], (err, response) => {
+				if (err) {
+					console.log('error on query ', err);
+				} else {
+					res.send(response.rows);
+				}
+			});
 		}
 	});
 });
-
 let port = process.env.PORT || 3010;
 app.listen(port, function() {
 	console.log(`listening on port ${port}`);
